ARC-441 fix: critical game table layout on mobile
